Magma Bubble Fix
Reverted teleport in temple on death from the previous fix because it ruins every screenshot taken on death by the client.
This new magma bubble code won't 'cause a death loop in your character and It will take only 10% of your health on step in (fire fields).
There is no better solution than this at the moment without ruin the client screenshots.