Fix frafnar npc
There is an error in the logic of the progress of this quest. The error was that the quest only progress if you joke about Bolfona's cake delivery, which is confusing and stuck players. Now progress unless you report about Bolfona delivery and it's optional to say if you mentioned him as the one who sent the cake to Bolfona.