Proposal: All-In-One OTServer NPC
1.- Deleted imbuement assistant and the lootmonger npcs.
2.- Added deleted NPC functions to new NPC trader djinn with real tibia prices for a fair game balance.
3.- Added commented item name in item.xml in case server owner wants to use.

How it works? If decided by server owner, you can use the djinn's lamp to enable the call shop command, this command can call a trader djinn to your position for 3 mintutes, it costs 100 soul so it doesn't break the game. It can be an useful feature for server or server shop.

Extra fix: I deleted the duplicate awarness of emperor, it has a wrong name and antoher PR already added the correct name to this NPC.